Post Experiment Commit
Multiple data recording methods were attempted at M City:
1. Using Polysync Scripts from Zexiang (now stored in polysync/polysync_utils/ directory)
2. MATLAB Script (polysync/record_data.m)
3. monitor.slx

Only monitor.slx worked. (It is the same one that Petter developed originally.)

The gains for the PID controllers in acc_pid_controller_yunus2.m were adjusted along with the LK pid controllers. (Not sure if those were also attached to this commit.)

The steering gains worked well in CarSim but we needed to remove the "I" and "D" terms when operating in M City.